pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "oss-issue-fixer-agent"
version = "0.1.0"
description = "Automated issue/feature fixing agent for OSS communities."
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0.1",
  "requests>=2.32.3",
]

[project.scripts]
oss-fixer = "oss_issue_fixer.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
